#938 - Refactor to remove RandomBuild.java
